#50 refactor: _event-listener(s) の命名・配置を一貫させる
- _event-listener → _event-listeners にリネーム。_components/_hooks/_stores 等、他の集約ディレクトリが複数形である点に揃えた
- ScopeEventListeners (_hooks/index.tsx) を _event-listeners/index.tsx へ移動。_hooks を汎用ラッパー (dispatcher/listener) のみの置き場に整理し、個別 listener 実装とそれを束ねる component を _event-listeners に集約した
