+++ src/prototypes/time-control/time-control-03/_events/_event-listeners/index.tsx
@@ -1,4 +1,4 @@
-import { useResetAllEventListener } from '../_event-listener/use-reset-all-event-listener'
+import { useResetAllEventListener } from './use-reset-all-event-listener'
 
 /**
  * scope 全体のイベント購読をまとめて有効化する
