+++ src/prototypes/time-control/time-control-03/_events/_event-listeners/use-reset-all-event-listener/index.ts
